+++ schemas/config/server.js
@@ -17,9 +17,9 @@
     timeout: 'number',
   },
   scheduler: {
-    concurrency: 'number',
-    size: 'number',
-    timeout: 'number',
+    concurrency: '?number',
+    size: '?number',
+    timeout: '?number',
   },
   workers: {
     pool: 'number',